Move util to common
