+++ src/main/comments.js
@@ -9,7 +9,7 @@ const indent = docBuilders.indent;
 const lineSuffix = docBuilders.lineSuffix;
 const join = docBuilders.join;
 const cursor = docBuilders.cursor;
-const util = require("../builder/util");
+const util = require("../common/util");
 const childNodesCacheKey = Symbol("child-nodes");
 const locStart = util.locStart;
 const locEnd = util.locEnd;
